requirements.txt
requests
bs4
tqdm
pandas
scikit-learn
numpy
matplotlib
seaborn
tensorflow
tensorflow-metal
keras
python-dotenv
opencv-python
supervision
inference[transformers,sam,sam3,clip,gaze,grounding-dino,yolo-world]
sam3
timm
torch
torchvision
pytorch-metric-learning
roboflow
supervision
samv2
streamlit
streamlit-drawable-canvas
annoy